use same base class for sheet music listing and other hymn listing; enable sorting and filtering of sheet music listing
